requirements.txt
bugsnag==3.6.0
celery==4.3.0
coreapi==2.3.3
coverage==4.5.3
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-ipware==2.1.0
django-suit==0.2.26
Django==2.2.2
djangorestframework-jwt==1.11.0
djangorestframework==3.9.4
Fabric3>=1.14
flake8==3.7.7
isort==4.3.21
python-decouple==3.1
redis==3.2.1
